Displayed decimal point percentages

+++ Assignment_1/src/com/Assignment_1/test/SentenceReader.java
@@ -12,9 +12,9 @@ public class SentenceReader
 {
 	//Attributes
 	int i = 0;
-	int wordcount = 0;
-	int formal = 0;
-	int slang = 0;
+	float wordcount = 0;
+	float formal = 0;
+	float slang = 0;
 	float total = 0;
 	float formalp = 0;
 	float slangp = 0;
